css move
Moved the css that was in the global.css file to the home.css. Later we
will need to sort out what code needs to be put in the global and which
needs put in the home for now we are going to keep it in the home.css
Also added the code in the index.html file to add home.css and also
fixed the file path for the header image in the home.css file.